Hata düzeltmeleri

+++ Yamala.sh
@@ -305,7 +305,12 @@ flatpak install -y flathub org.mozilla.firefox
 echo "# Yerel yazılımlar yükleniyor." ; sleep 2
 
 apt-get install -y gir1.2-flatpak-1.0
+
+if [ ! -d usr/share/pardusyama/ ]; then
 dpkg -R --install ./Yazılımlar/
+else
+dpkg -R --install usr/share/pardusyama/Yazılımlar/
+fi
 apt-get -fy install
 
 
@@ -322,11 +327,21 @@ echo "# Oyuncu araçları yükleniyor." ; sleep 2
 
 # flatpak install flathub com.valvesoftware.Steam  #Flatpak runtime sürümü depoda eski, bu eski sürümde de Nvidia-Steam sorunlu. O yüzden depoda Flatpak güncellenene kadar deb sürümü yükleyeceğiz. 
 
+if [ ! -d usr/share/pardusyama/ ]; then
 cd ./Oyuncu
+else
+cd usr/share/pardusyama/Oyuncu
+fi
+
 wget https://steamcdn-a.akamaihd.net/client/installer/steam.deb 2>&1 | sed -u 's/.* \([0-9]\+%\)\ \+\([0-9.]\+.\) \(.*\)/\1\n# İndirme Hızı \2\/s, Kalan zaman \3/' | zenity --progress --title "İndirme İşlemi" --text "Steam indiriliyor..." --no-cancel --auto-close --pulsate
 chmod 777 steam*.deb
 cd ..
+
+if [ ! -d usr/share/pardusyama/ ]; then
 dpkg -R --install ./Oyuncu/
+else
+dpkg -R --install usr/share/pardusyama/Oyuncu/
+fi
 apt-get -fy install
 
 #lutris --reinstall epic-games-store
@@ -370,7 +385,13 @@ groupadd smbgrp
 
 usermod ${u} -aG smbgrp
 mv /etc/samba/smb.conf /etc/samba/defsmb.conf
+
+if [ ! -d usr/share/pardusyama/ ]; then
 cp -r ./smb.conf /etc/samba/
+else
+cp -r usr/share/pardusyama/smb.conf /etc/samba/
+fi
+
 find "/var/lib/samba/usershares" -type f -exec chmod 777 {} \+ # Samba izinleri. Şimdilik böyle. Alternatif çözüm üretilince değiştilecek.
 chown -R $(id -un ${u}):$(id -gn ${u}) "/var/lib/samba/usershares"
 
@@ -384,10 +405,16 @@ systemctl restart smbd.service
 echo "# Sağ Tık menüsü geliştiriliyor..." ; sleep 2
 
 SAB="Şablonlar"
-BET=".local/share/nautilus/scripts"               
+BET=".local/share/nautilus/scripts"      
+
+if [ ! -d usr/share/pardusyama/Oyuncu/ ];
+then
 _FILESB="./Betikler/*"	
 _FILESS="./Şablonlar/*"  
-
+else      
+_FILESB="usr/share/pardusyama/Betikler/*"	
+_FILESS="usr/share/pardusyama/Şablonlar/*"  
+fi
   
   
   # .config/user-dirs.dirs dosyası yoksa oluştur.
@@ -441,15 +468,21 @@ killall xfconfd
 
 #dpkg -R --install ./Xfce/deb
 #apt-get -fy install
-
+if [ ! -d usr/share/pardusyama/ ]; then
 cp -r ./Tema/Genel/* /usr/share/themes/
 cp -r ./Tema/Simge/* /usr/share/icons/
+else
+cp -r usr/share/pardusyama/Tema/Genel/* /usr/share/themes/
+cp -r usr/share/pardusyama/Tema/Simge/* /usr/share/icons/
+fi
 gtk-update-icon-cache	"/usr/share/icons/Qogir/"							 	 # Simge ön belleğini temizle
 gtk-update-icon-cache	"/usr/share/icons/Qogir-dark/"							 # Simge ön belleğini temizle
 
-
+if [ ! -d usr/share/pardusyama/ ]; then
 _FILESX="./Xfce/.config/."
-
+else
+_FILESX="usr/share/pardusyama/Xfce/.config/."
+fi
   
 
    for f in $_FILESX
@@ -558,9 +591,13 @@ GNOME_SITE="https://extensions.gnome.org/extension-data/"
 
 #unzip -c <extension zip file name> metadata.json | grep uuid | cut -d \" -f4
 
-GNM=".local/share/gnome-shell/extensions" 		
-_FILESG="./Gnome/*"
+GNM=".local/share/gnome-shell/extensions" 	
 
+if [ ! -d usr/share/pardusyama/ ]; then	
+_FILESG="./Gnome/*"
+else
+_FILESG="usr/share/pardusyama/Gnome/*"
+fi
 
    for f in $_FILESG
    do
@@ -582,8 +619,14 @@ done
 
 echo "# Sistem ince ayarları yapılıyor." ; sleep 2
 
+if [ ! -d usr/share/pardusyama/ ]; then	
 cp -r ./Tema/Genel/* /usr/share/themes/
 cp -r ./Tema/Simge/* /usr/share/icons/
+else
+cp -r usr/share/pardusyama/Tema/Genel/* /usr/share/themes/
+cp -r usr/share/pardusyama/Tema/Simge/* /usr/share/icons/
+fi
+
 gtk-update-icon-cache	"/usr/share/icons/Qogir"							 	 # Simge ön belleğini temizle
 gtk-update-icon-cache	"/usr/share/icons/Qogir-dark"							 # Simge ön belleğini temizle
 
@@ -701,16 +744,10 @@ dconf update
 
 echo "# Windows fontları yükleniyor." ; sleep 2
 
-
-	# /usr/share/fonts/truetype/msttcorefonts kasörü var mı kontrol et
-	if [ ! -d /usr/share/fonts/truetype/msttcorefonts ];
-	then
-		mkdir /usr/share/fonts/truetype/msttcorefonts
-		cp -r Fontlar/* /usr/share/fonts/truetype/msttcorefonts
-		echo -e "\Fontlar yüklendi.."
-	else
-		# varsa yeni klasör oluşturmayı atla
-		cp -r Fontlar/* /usr/share/fonts/truetype/msttcorefonts
+		if [ ! -d usr/share/pardusyama/ ]; then	
+		cp -r ./Fontlar/* /usr/share/fonts/truetype/msttcorefonts
+else
+		cp -r usr/share/pardusyama/Fontlar/* /usr/share/fonts/truetype/msttcorefonts
 		echo "\Ücretsiz Windows Fontları yüklendi."
 	fi
 
@@ -736,7 +773,11 @@ GFXBT=4096x2160,1920x1080,1366x768,1024x768,auto
 
   # Copy theme
 
+if [ ! -d usr/share/pardusyama/ ]; then	
   cp -a ./Tema/Grub/${THEME_NAME}/* ${THEME_DIR}/${THEME_NAME}
+else
+	cp -a usr/share/pardusyama/Tema/Grub/${THEME_NAME}/* ${THEME_DIR}/${THEME_NAME}
+fi
 
   # Set theme
